move workgroup package code into concurrency package
